requirements.txt
fastapi
uvicorn
pydantic
redis
langchain
langchain-core
langchain-openai
openpyxl
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-multipart
python-jose[cryptography]